+++ packages/evm/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @boostxyz/evm
 
+## 6.0.3
+
+### Patch Changes
+
+- df2369d: export LimitedSignerValidator component interface
+
+## 6.0.2
+
+### Patch Changes
+
+- 8c3369c: Deployed LimitedSignerValidator
+
 ## 6.0.1
 
 ### Patch Changes